Set ContentLength = 0 for requests with no body
The remote server responds with error 411 Length Required when posting a request with no message-body. 

See: https://www.w3.org/Protocols/rfc2616/rfc2616-sec10.html

10.4.12 411 Length Required

The server refuses to accept the request without a defined Content- Length. The client MAY repeat the request if it adds a valid Content-Length header field containing the length of the message-body in the request message.