Set ContentLength = 0 for requests with no body
The remote server responds with error 411 Length Required when posting a request with no message-body. Setting ContentLength = 0 for the request fixes it.

I also moved the outputStream.Write code into the if block so it doesn't throw a null reference exception (which was not being handled by the using statement.