Fix duplication of headers containing more than one :
The previous behaviour where `$pos` matched on the last occurrence of a colon was breaking for headers like `Expires` that contain colons in the value of the header. As such we were passing a value like "Expires: Tue, 27 Sep 2016 09:" to header_remove(), which resutled in the header not actually being removed.

This caused duplication of headers later when getResponse() added another Expires header.